More source format changes
Changing the name of the folder that sources are deployed to from
App_Packages to lib, because of hitting some name length limitations.
This still doesn't really solve the issue (which would really need to be
fixed inside visual studio in this case) but it does help a little to
minimize how often it will be hit.
bumping version number